refactor(node-wrapper): add separate folder
