+++ packages/react/src/components/NodeWrapper/utils.tsx
@@ -1,0 +1,8 @@
+import { XYPosition } from '@xyflow/system';
+
+export const arrowKeyDiffs: Record<string, XYPosition> = {
+  ArrowUp: { x: 0, y: -1 },
+  ArrowDown: { x: 0, y: 1 },
+  ArrowLeft: { x: -1, y: 0 },
+  ArrowRight: { x: 1, y: 0 },
+};
